requirements.txt
openai>=2.9.0
anthropic>=0.39.0
numpy>=1.26.0
scipy>=1.13.0
pyyaml>=6.0.2
matplotlib>=3.9.0
seaborn>=0.13.2
tiktoken>=0.8.0

# CLI tool dependencies
click>=8.1.0
ruamel.yaml>=0.18.0
rich>=13.0.0
pydantic>=2.0.0

